setIntervalの処理を統合
