+++ app/javascript/packs/application.js
@@ -12,7 +12,6 @@ require("../canvas")
 require("../form")
 require("../drawing")
 require("../review_post")
-require("../review_delete")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
